refactor: update CloudWatchMetrics directory and component to reflect new naming conventions
