+++ client/src/pages/FunctionAnalytics/Throttle/Throttle.tsx
