+++ client/src/pages/FunctionAnalytics/TotalDuration/TotalDuration.tsx
